+++ client/src/api/services/posts.ts
@@ -1,15 +1,13 @@
-import { api, axiosInstance } from "@/api/instance";
+import { axiosInstance } from "@/api/instance";
 import { InfiniteScrollResponse, LatestPostsApiResponse, TrendingPostsApiResponse, Post } from "@/types/post";
-
-const instance = import.meta.env.DEV ? api : axiosInstance;
 
 export const posts = {
   trending: async (): Promise<TrendingPostsApiResponse> => {
-    const response = await instance.get<TrendingPostsApiResponse>("/api/feed/trend");
+    const response = await axiosInstance.get<TrendingPostsApiResponse>("/api/feed/trend");
     return response.data;
   },
   latest: async (params: { limit: number; lastId: number }): Promise<InfiniteScrollResponse<Post>> => {
-    const response = await instance.get<LatestPostsApiResponse>("/api/feed", {
+    const response = await axiosInstance.get<LatestPostsApiResponse>("/api/feed", {
       params: {
         limit: params.limit,
         lastId: params.lastId || 0,
